Refactor architecture into modular package
