+++ architecture/__init__.py
@@ -1,0 +1,62 @@
+from .llm import (
+    PROMPT_PACKET,
+    VALIDATION_OUTPUT_SCHEMA,
+    SCOUT_OUTPUT_SCHEMA,
+    LLM_FAILURES,
+    call_llm_with_schema,
+)
+from .models import (
+    Status,
+    PlanKind,
+    CodeSpan,
+    Evidence,
+    SustainingCondition,
+    ValidationInput,
+    ValidationResult,
+    Finding,
+    AgentReport,
+    AuditReport,
+    render_report,
+    _id,
+)
+from .adapters import CodebaseAdapter, WebSearchAdapter, LLMClient
+from .strategies import (
+    ValidationStrategy,
+    StrategyRegistry,
+    PathExistsStrategy,
+    IsUserControlledStrategy,
+)
+from .agents import ValidationAgent, TinyShellAgent
+from .orchestrator import OrchestratorConfig, Task, AuditOrchestrator
+
+__all__ = [
+    "PROMPT_PACKET",
+    "VALIDATION_OUTPUT_SCHEMA",
+    "SCOUT_OUTPUT_SCHEMA",
+    "LLM_FAILURES",
+    "call_llm_with_schema",
+    "Status",
+    "PlanKind",
+    "CodeSpan",
+    "Evidence",
+    "SustainingCondition",
+    "ValidationInput",
+    "ValidationResult",
+    "Finding",
+    "AgentReport",
+    "AuditReport",
+    "render_report",
+    "_id",
+    "CodebaseAdapter",
+    "WebSearchAdapter",
+    "LLMClient",
+    "ValidationStrategy",
+    "StrategyRegistry",
+    "PathExistsStrategy",
+    "IsUserControlledStrategy",
+    "ValidationAgent",
+    "TinyShellAgent",
+    "OrchestratorConfig",
+    "Task",
+    "AuditOrchestrator",
+]
